DBMA-258  Set up system which gets user episode data from API also set up system which updates the stop time of episodes for users and applies that to the AudioPlayer

+++ DABApp/DABApp/DabData/dbPlayerActions.cs
@@ -12,7 +12,7 @@ namespace DABApp
 		public string entity_type { get; set;}
 		public int EpisodeId { get; set;}
 		public string ActionType { get; set;}
-		public decimal PlayerTime { get; set;}
+		public double PlayerTime { get; set;}
 		public string UserEmail { get; set;}
 	}
 }
